move health to common
