+++ byteguide/__init__.py
@@ -10,7 +10,6 @@ from byteguide.config import config
 from byteguide.libs.jinja_fltrs import register_filters
 from byteguide.routes.common import common_routes
 from byteguide.routes.display import display_routes
-from byteguide.routes.health import health_routes
 from byteguide.routes.manage import manage_routes
 
 
@@ -34,7 +33,6 @@ def register_blueprints(application):
     application.register_blueprint(common_routes)
     application.register_blueprint(display_routes)
     application.register_blueprint(manage_routes)
-    application.register_blueprint(health_routes)
 
 
 def create_symlink():
@@ -49,10 +47,14 @@ def create_symlink():
     # a symlink to the static folder for simplicity of serving files
     if not config.docfiles_dir.startswith("byteguide/static"):
         try:
-            log.info(f"Creating symlink from {docfiles_dir} to byteguide/static/docfiles")
+            log.info(
+                f"Creating symlink from {docfiles_dir} to byteguide/static/docfiles"
+            )
             # Create symlink if it does not already exist
             if not os.path.exists("byteguide/static/docfiles"):
-                os.symlink(docfiles_dir, "byteguide/static/docfiles", target_is_directory=True)
+                os.symlink(
+                    docfiles_dir, "byteguide/static/docfiles", target_is_directory=True
+                )
         except OSError as e:
             # Handle symlink creation failure
             log.error(f"Symlink creation failed: {e}")
